Feat: 사용자 정보 조회, 닉네임/재화/소유색상 변경 API 추가
- 로딩 씬에서 사용자 닉네임, 적용 색상, 재화량, 소유 색상 리스트 반환 가능한 API
- 현재 적용 색상 변경 API
- 색상 구매 시 호출 할 API
- 닉네임 또는 쿠도스 변경 API

+++ Web Server/src/main/java/com/example/fallguys/dto/user/UserResponseDto.java
@@ -8,7 +8,7 @@ public class UserResponseDto {
     private final String userId;
     private final String userNickname;
     private final int userKudos;
-    private final int userCostumeColor;
+    private final Long userCostumeColor;
 
     public UserResponseDto(User user) {
         this.userId = user.getUserId();
